Added in console.table npm, and have both the Add Employee and Exit functions working correctly.

+++ employees.sql
@@ -3,16 +3,16 @@ CREATE DATABASE employees_db;
 USE employees_db;
 
 CREATE TABLE departments(
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
-  name VARCHAR(30)
+  id INTEGER AUTO_INCREMENT NOT NULL,
+  name VARCHAR(30),
   PRIMARY KEY (id)
 );
 
 CREATE TABLE role(
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
+  id INTEGER AUTO_INCREMENT NOT NULL,
   title VARCHAR(30),
   salary DECIMAL(12,2),
-  department_id INTEGER(11)
+  department_id INTEGER,
   PRIMARY KEY (id)
 );
 
@@ -20,7 +20,7 @@ CREATE TABLE employee(
   id INTEGER(11) AUTO_INCREMENT NOT NULL,
   first_name VARCHAR(30),
   last_name VARCHAR(30),
-  role_id INTEGER(11),
-  manager_id INTEGER(11) NULL,
+  role_id INTEGER,
+  manager_id INTEGER NULL,
   PRIMARY KEY (id)
 );
